+++ src/Fulma.Elmish/RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 0.1.0-beta-002
+
+* `DatePicker`: It can be open even if the input is already focused
+* `DatePicker`: Allow users to customize the position of the "calendar box" via the config
+
 ### 0.1.0-beta-001
 
 * Init project